feat: adicionar alteracoes da branch dev

+++ src/services/student.services.ts
@@ -15,6 +15,7 @@ import {
 import { IVerifyUpdateByCategory } from '../intefaces/verify.interfaces';
 import { LoginDALs } from '../database/repositories/user.repositories/user.dals/login.dals';
 import { hash } from 'bcrypt';
+import { PictureDALs } from '../database/repositories/user.repositories/user.dals/picture.dals';
 
 class StudentService {
   private readonly studentDALs: StudentDALs;
@@ -26,6 +27,7 @@ class StudentService {
   private readonly typeGrantDALs: TypeGrantDALs;
   private readonly userDALs: UserDALs;
   private readonly enrollmentDALs: EnrollmentDALs;
+  private readonly pictureDALs: PictureDALs;
 
   constructor() {
     this.studentDALs = new StudentDALs();
@@ -37,6 +39,7 @@ class StudentService {
     this.visitorDALs = new VisitorDALs();
     this.employeeDALs = new EmployeeDALs();
     this.enrollmentDALs = new EnrollmentDALs();
+    this.pictureDALs = new PictureDALs();
   }
 
   async createStudent({
@@ -72,7 +75,7 @@ class StudentService {
 
     if (loginByEmail) {
       throw new BadRequestError({
-        message: 'email already exists, only one email is allowed.',
+        message: 'Email already exists, only one email is allowed.',
       });
     }
     const passwordHash = await hash(password, 10);
@@ -103,7 +106,16 @@ class StudentService {
       dailyMeals: dailyMeals,
       loginUserId: createLogin.id,
     });
+     let url = "";
 
+    if (picture) {
+      url = picture;
+    }
+
+    const createPicture = await this.pictureDALs.createPicture({
+      url: url,
+      userId: createUser.id,
+    });
     if (!enrollment) {
       throw new BadRequestError({ message: 'Enrollment is required' });
     }
@@ -120,6 +132,7 @@ class StudentService {
       personName: createPerson.name,
       categoryName: getCategory.name,
       typeGrantName: getTypeGrant.name,
+      picture: createPicture.url,
       dailyMeals: dailyMeals,
       loginData: {
         email: createLogin.email,
@@ -137,22 +150,28 @@ class StudentService {
     if (!enrollment) {
       throw new BadRequestError({ message: 'Enrollment is required' });
     }
-   
+
     switch (oldCategory) {
       case 'FUNCIONARIO':
-        const oldEnrollmentEmployee = await this.employeeDALs.findEnrrolmentByUserId(
-          userId,
-        );
-         if (!oldEnrollmentEmployee) {
+        const oldEnrollmentEmployee =
+          await this.employeeDALs.findEnrrolmentByUserId(userId);
+        if (!oldEnrollmentEmployee) {
           throw new BadRequestError({ message: 'OldEnrollment is required' });
         }
-        
-        if(oldEnrollmentEmployee.enrollment === enrollment){
-             throw new UnprocessedEntityError({
+
+        if (oldEnrollmentEmployee.enrollment === enrollment) {
+          throw new UnprocessedEntityError({
             message: 'category cannot be update without enrrolment',
+          });
+        }
+
+        const IsEnrrolmentUnique = this.enrollmentDALs.checkEnrollmentUnique(enrollment);
+        if(!IsEnrrolmentUnique){
+          throw new BadRequestError({
+            message: 'Enrollment already exists, only one enrollment is allowed.',
           }); 
         }
-        
+
         const employee = await this.employeeDALs.deleteByUserId(userId);
         const updateEnrollment = await this.enrollmentDALs.updateEnrollment(
           employee.enrollmentId,
@@ -163,6 +182,7 @@ class StudentService {
           enrollmentId: updateEnrollment.id,
         });
       case 'VISITANTE':
+        
         await this.visitorDALs.deleteByUserId(userId);
         const createEnrollment = await this.enrollmentDALs.createEnrollment(
           enrollment,
@@ -173,23 +193,22 @@ class StudentService {
         });
 
       case 'ESTUDANTE':
-        const oldEnrollmentStudent = await this.studentDALs.findEnrrolmentByUserId(
-          userId,
-        );
-        
-         if (!oldEnrollmentStudent) {
+        const oldEnrollmentStudent =
+          await this.studentDALs.findEnrrolmentByUserId(userId);
+
+        if (!oldEnrollmentStudent) {
           throw new BadRequestError({ message: 'OldEnrollment is required' });
         }
-      
-       if (
-            oldCategory !== category || oldEnrollmentStudent.enrollment !== enrollment
+
+        if (
+          oldCategory !== category ||
+          oldEnrollmentStudent.enrollment !== enrollment
         ) {
           throw new UnprocessedEntityError({
             message: 'enrrolment cannot be update without category',
           });
-          
         }
-        return await this.studentDALs.findStudentByUserId( userId );
+        return await this.studentDALs.findStudentByUserId(userId);
       default:
         throw new BadRequestError({ message: 'Old Category NotFound' });
     }
